requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
qrcode==7.3.1
Pillow==8.3.1
python-dotenv==0.19.0
gunicorn==20.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==0.15.1
email-validator==1.1.3
Werkzeug==2.3.7 